Reworked the processor to not block on init/start (#12)
The initialize, start, stop are now based on a loose state machines of when to send them during message processing. This keeps the init/start/stop from going out of order, as well as blocking the spawn functions. Additionally this fixes an issue with a panic in start/init actually not being captured properly by the restart mechanisms when spawning a new proc. Closes #11 by no longer failing when calling Start on a process.